I have fixed the scale change bug with a work around
Forst I changed everything to scale 1, this introduced clipping plane issues while riding train forcing me to put the transform above the train and not on front of it. I ended up just using a camera and not deactivating the VR rig. Instead I disabled the hands while riding the rain, re-enabling when getting off.